Fix packet_utils.py checksum in python3
Python2's array.array('H', x) does not work as expected if x is of type
bytearray, but it does work if x is of type str. In Python3 this works
regardless of whether x is bytes or bytearray. bytes(x) evaluates to
str(x) in py2 and bytes(x) in py3, so it works for both.

Signed-off-by: James Guthrie <james@prodigi.ch>